Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
google-auth==2.36.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
curl_cffi==0.13.0
fastapi==0.135.1
frozendict==2.4.7
h11==0.16.0
idna==3.11
joblib==1.5.3
lxml==6.0.2
multitasking==0.0.12
numpy==2.4.3
pandas==3.0.1
passlib==1.7.4
peewee==4.0.2
platformdirs==4.9.4
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
tfinance==0.1.4
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
xgboost==3.2.0
yfinance==1.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
